+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - transition between collection and viewer when cutout area is not used
+- saving video playback state when leaving viewer
 
 ## <a id="v1.7.8"></a>[v1.7.8] - 2022-12-20